requirements.txt
celery
celery[redis]
flask
requests
pymongo
redis
pyyaml
configobj